pyproject.toml
[project]
name = "day-01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.0.1",
    "fastapi>=0.135.1",
    "passlib[bcrypt]>=1.7.4",
    "prisma>=0.15.0",
    "pydantic[email]>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn[standard]>=0.41.0",
]
